Added menu continue game button, added game ending functionallity for round mode and point mode

+++ Quartett/app/src/main/java/de/in/uulm/map/quartett/gamesettings/GameSettingsContract.java
@@ -25,7 +25,7 @@ public interface GameSettingsContract {
 
         int getPoints();
 
-        int getTime();
+        long getTime();
 
         int getRounds();
